updated layouts and header
