+++ app/[locale]/account/brews/[brew_id]/layout.tsx
@@ -9,11 +9,9 @@ import { ReactNode } from "react";
 
 function BrewLayout({ children }: { children: ReactNode }) {
   return (
-    <div className="w-11/12 max-w-[1200px] relative rounded-xl bg-background px-4 py-6 sm:px-12 sm:py-8">
-      <RecipeProvider>
-        <Nutrients>{children}</Nutrients>
-      </RecipeProvider>
-    </div>
+    <RecipeProvider>
+      <Nutrients>{children}</Nutrients>
+    </RecipeProvider>
   );
 }
