+++ app/[locale]/account/brews/layout.tsx
@@ -1,0 +1,13 @@
+import Header from "@/components/account/header";
+import { ReactNode } from "react";
+
+function Layout({ children }: { children: ReactNode }) {
+  return (
+    <div className="w-11/12 max-w-[1200px] relative rounded-xl bg-background px-4 py-6 sm:px-12 sm:py-8">
+      <Header />
+      {children}
+    </div>
+  );
+}
+
+export default Layout;
