+++ app/[locale]/account/brews/new/page.tsx
@@ -78,146 +78,138 @@ export default function NewBrewClient() {
   if (isLoading) return <NewBrewSkeleton />;
 
   return (
-    <div className="w-11/12 max-w-[1200px] relative rounded-xl bg-background px-4 py-6 sm:px-12 sm:py-8">
-      <div className="my-6 space-y-4">
-        <div className="flex items-center justify-between gap-3">
-          <h2 className="text-2xl">{t("brews.new", "New Brew")}</h2>
+    <div className="my-6 space-y-4">
+      <div className="flex items-center justify-between gap-3">
+        <h2 className="text-2xl">{t("brews.new", "New Brew")}</h2>
 
-          <Button variant="secondary" asChild>
-            <Link href="/account/brews">{t("back", "Back")}</Link>
-          </Button>
-        </div>
+        <Button variant="secondary" asChild>
+          <Link href="/account/brews">{t("back", "Back")}</Link>
+        </Button>
+      </div>
 
-        {/* Optional: basic setup fields */}
-        <div className="rounded-xl border border-border bg-card p-4 space-y-3">
-          <div className="text-sm text-muted-foreground">
-            {t(
-              "brews.new.help",
-              "Pick a recipe to start a brew. (Optional) set a name and starting volume."
-            )}
-          </div>
-
-          <div className="grid gap-3 sm:grid-cols-2">
-            <div className="space-y-1">
-              <div className="text-sm font-medium">{t("name", "Name")}</div>
-              <Input
-                value={nameDraft}
-                onChange={(e) => setNameDraft(e.target.value)}
-                placeholder={t(
-                  "brews.new.namePlaceholder",
-                  "e.g. January Batch"
-                )}
-                disabled={createMutation.isPending}
-              />
-            </div>
-
-            <div className="space-y-1">
-              <div className="text-sm font-medium">
-                {t("brews.new.volumeLiters", "Starting volume (L)")}
-              </div>
-              <Input
-                value={volumeDraft}
-                onChange={(e) => setVolumeDraft(e.target.value)}
-                inputMode="decimal"
-                placeholder="—"
-                disabled={createMutation.isPending}
-              />
-            </div>
-          </div>
-        </div>
-
-        {/* Search */}
-        <div className="flex flex-wrap items-center gap-3 justify-between">
-          <InputGroup className="w-full sm:max-w-sm">
-            <InputGroupInput
-              value={searchValue}
-              onChange={(e) => search(e.target.value)}
-              placeholder={t("search", "Search")}
-              disabled={createMutation.isPending}
-            />
-            <InputGroupAddon>
-              <Search />
-            </InputGroupAddon>
-            <InputGroupAddon align="inline-end">
-              <InputGroupButton
-                title={t("clear", "Clear")}
-                onClick={clearSearch}
-                className={cn({ hidden: searchValue.length === 0 })}
-                disabled={createMutation.isPending}
-              >
-                <X />
-              </InputGroupButton>
-            </InputGroupAddon>
-          </InputGroup>
-
-          {/* page size (you can reuse your Select UI if you want; keeping it minimal) */}
-        </div>
-
-        {/* Cards */}
-        <div className="flex flex-wrap justify-center gap-4 py-2">
-          {pageData.length > 0 ? (
-            pageData.map((rec) => (
-              <RecipeStartCard
-                key={rec.id}
-                recipe={rec}
-                isCreating={createMutation.isPending}
-                onStart={async () => {
-                  try {
-                    const vol =
-                      volumeDraft.trim().length === 0
-                        ? undefined
-                        : Number(volumeDraft);
-
-                    if (vol !== undefined && !Number.isFinite(vol)) {
-                      toast({
-                        description: t(
-                          "error.invalidNumber",
-                          "Invalid number."
-                        ),
-                        variant: "destructive"
-                      });
-                      return;
-                    }
-
-                    const brew = await createMutation.mutateAsync({
-                      recipe_id: rec.id,
-                      name: nameDraft.trim() ? nameDraft.trim() : undefined,
-                      current_volume_liters: vol
-                    });
-
-                    router.push(`/account/brews/${brew.id}`);
-                  } catch (e) {
-                    console.error(e);
-                    toast({
-                      description: t("error.generic", "Something went wrong."),
-                      variant: "destructive"
-                    });
-                  }
-                }}
-              />
-            ))
-          ) : (
-            <p className="w-full text-center mt-6">
-              {filteredData.length === 0
-                ? t("noResults", "No results found.")
-                : t("accountPage.noRecipes", "No recipes found.")}
-            </p>
+      {/* Optional: basic setup fields */}
+      <div className="rounded-xl border border-border bg-card p-4 space-y-3">
+        <div className="text-sm text-muted-foreground">
+          {t(
+            "brews.new.help",
+            "Pick a recipe to start a brew. (Optional) set a name and starting volume."
           )}
         </div>
 
-        {/* Pagination */}
-        {filteredData.length > 0 && totalPages > 1 && (
-          <AccountPagination
-            page={page}
-            totalPages={totalPages}
-            canPrev={page > 1}
-            canNext={page < totalPages}
-            onPrev={prevPage}
-            onNext={nextPage}
-            onGoTo={goToPage}
+        <div className="grid gap-3 sm:grid-cols-2">
+          <div className="space-y-1">
+            <div className="text-sm font-medium">{t("name", "Name")}</div>
+            <Input
+              value={nameDraft}
+              onChange={(e) => setNameDraft(e.target.value)}
+              placeholder={t("brews.new.namePlaceholder", "e.g. January Batch")}
+              disabled={createMutation.isPending}
+            />
+          </div>
+
+          <div className="space-y-1">
+            <div className="text-sm font-medium">
+              {t("brews.new.volumeLiters", "Starting volume (L)")}
+            </div>
+            <Input
+              value={volumeDraft}
+              onChange={(e) => setVolumeDraft(e.target.value)}
+              inputMode="decimal"
+              placeholder="—"
+              disabled={createMutation.isPending}
+            />
+          </div>
+        </div>
+      </div>
+
+      {/* Search */}
+      <div className="flex flex-wrap items-center gap-3 justify-between">
+        <InputGroup className="w-full sm:max-w-sm">
+          <InputGroupInput
+            value={searchValue}
+            onChange={(e) => search(e.target.value)}
+            placeholder={t("search", "Search")}
+            disabled={createMutation.isPending}
           />
+          <InputGroupAddon>
+            <Search />
+          </InputGroupAddon>
+          <InputGroupAddon align="inline-end">
+            <InputGroupButton
+              title={t("clear", "Clear")}
+              onClick={clearSearch}
+              className={cn({ hidden: searchValue.length === 0 })}
+              disabled={createMutation.isPending}
+            >
+              <X />
+            </InputGroupButton>
+          </InputGroupAddon>
+        </InputGroup>
+
+        {/* page size (you can reuse your Select UI if you want; keeping it minimal) */}
+      </div>
+
+      {/* Cards */}
+      <div className="flex flex-wrap justify-center gap-4 py-2">
+        {pageData.length > 0 ? (
+          pageData.map((rec) => (
+            <RecipeStartCard
+              key={rec.id}
+              recipe={rec}
+              isCreating={createMutation.isPending}
+              onStart={async () => {
+                try {
+                  const vol =
+                    volumeDraft.trim().length === 0
+                      ? undefined
+                      : Number(volumeDraft);
+
+                  if (vol !== undefined && !Number.isFinite(vol)) {
+                    toast({
+                      description: t("error.invalidNumber", "Invalid number."),
+                      variant: "destructive"
+                    });
+                    return;
+                  }
+
+                  const brew = await createMutation.mutateAsync({
+                    recipe_id: rec.id,
+                    name: nameDraft.trim() ? nameDraft.trim() : undefined,
+                    current_volume_liters: vol
+                  });
+
+                  router.push(`/account/brews/${brew.id}`);
+                } catch (e) {
+                  console.error(e);
+                  toast({
+                    description: t("error.generic", "Something went wrong."),
+                    variant: "destructive"
+                  });
+                }
+              }}
+            />
+          ))
+        ) : (
+          <p className="w-full text-center mt-6">
+            {filteredData.length === 0
+              ? t("noResults", "No results found.")
+              : t("accountPage.noRecipes", "No recipes found.")}
+          </p>
         )}
       </div>
+
+      {/* Pagination */}
+      {filteredData.length > 0 && totalPages > 1 && (
+        <AccountPagination
+          page={page}
+          totalPages={totalPages}
+          canPrev={page > 1}
+          canNext={page < totalPages}
+          onPrev={prevPage}
+          onNext={nextPage}
+          onGoTo={goToPage}
+        />
+      )}
     </div>
   );
 }
